+++ apps/platform/src/app/(main)/page.tsx
@@ -6,11 +6,11 @@ import { toast } from 'sonner'
 import { useAtom, useAtomValue, useSetAtom } from 'jotai'
 import ProjectCard from '@/components/dashboard/project/projectCard'
 import ControllerInstance from '@/lib/controller-instance'
-import ProjectScreenLoader from '@/components/ui/project-screen-loader'
+import ProjectScreenLoader from '@/components/dashboard/project/projectScreenLoader'
 import CreateProjectDialogue from '@/components/dashboard/project/createProjectDialogue'
 import {
-  createProjectDialogOpenAtom,
-  currentWorkspaceAtom,
+  createProjectOpenAtom,
+  selectedWorkspaceAtom,
   projectsOfWorkspaceAtom
 } from '@/store'
 import EditProjectSheet from '@/components/dashboard/project/editProjectSheet'
@@ -19,8 +19,8 @@ import { Button } from '@/components/ui/button'
 export default function Index(): JSX.Element {
   const [loading, setLoading] = useState<boolean>(false)
 
-  const setIsCreateProjectDialogOpen = useSetAtom(createProjectDialogOpenAtom)
-  const currentWorkspace = useAtomValue(currentWorkspaceAtom)
+  const setIsCreateProjectDialogOpen = useSetAtom(createProjectOpenAtom)
+  const selectedWorkspace = useAtomValue(selectedWorkspaceAtom)
 
   // Projects to be displayed in the dashboard
   const [projects, setProjects] = useAtom(projectsOfWorkspaceAtom)
@@ -32,19 +32,24 @@ export default function Index(): JSX.Element {
     async function getAllProjects() {
       setLoading(true)
 
-      if (currentWorkspace) {
+      if (selectedWorkspace) {
         const { success, error, data } =
           await ControllerInstance.getInstance().projectController.getAllProjects(
-            { workspaceSlug: currentWorkspace.slug },
+            { workspaceSlug: selectedWorkspace.slug },
             {}
           )
 
         if (success && data) {
           setProjects(data.items)
         } else {
-          toast.error(
-            'Something went wrong while fetching projects. Check console for more info.'
-          )
+          toast.error('Something went wrong!', {
+            description: (
+              <p className="text-xs text-red-300">
+                Something went wrong while fetching projects. Check console for
+                more info.
+              </p>
+            )
+          })
           // eslint-disable-next-line no-console -- we need to log the error
           console.error(error)
         }
@@ -54,7 +59,7 @@ export default function Index(): JSX.Element {
     }
 
     getAllProjects()
-  }, [currentWorkspace, setProjects])
+  }, [selectedWorkspace, setProjects])
 
   return (
     <div className="flex flex-col gap-4">
