+++ apps/platform/src/app/(main)/project/[project]/@secret/page.tsx
@@ -1,54 +1,52 @@
 'use client'
 import React, { useEffect, useState } from 'react'
-import { usePathname } from 'next/navigation'
-import dayjs, { extend } from 'dayjs'
+import { extend } from 'dayjs'
 import relativeTime from 'dayjs/plugin/relativeTime'
-import { NoteIconSVG } from '@public/svg/secret'
+import { SecretLogoSVG } from '@public/svg/secret'
 import type { GetAllSecretsOfProjectResponse } from '@keyshade/schema'
-import {
-  Accordion,
-  AccordionContent,
-  AccordionItem,
-  AccordionTrigger
-} from '@/components/ui/accordion'
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow
-} from '@/components/ui/table'
+import { useAtomValue, useSetAtom } from 'jotai'
+import { toast } from 'sonner'
+import { Accordion } from '@/components/ui/accordion'
 import { ScrollArea } from '@/components/ui/scroll-area'
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger
-} from '@/components/ui/tooltip'
-import { Skeleton } from '@/components/ui/skeleton'
 import ControllerInstance from '@/lib/controller-instance'
+import SecretLoader from '@/components/dashboard/secret/secretLoader'
+import { Button } from '@/components/ui/button'
+import { createSecretOpenAtom, selectedProjectAtom } from '@/store'
+import SecretCard from '@/components/dashboard/secret/secretCard'
 
 extend(relativeTime)
 
 function SecretPage(): React.JSX.Element {
-  const [allSecrets, setAllSecrets] =
-    useState<GetAllSecretsOfProjectResponse['items']>()
+  const setIsCreateSecretOpen = useSetAtom(createSecretOpenAtom)
+  const [secrets, setSecrets] = useState<
+    GetAllSecretsOfProjectResponse['items']
+  >([])
   const [isLoading, setIsLoading] = useState<boolean>(true)
-  const pathname = usePathname()
+  const selectedProject = useAtomValue(selectedProjectAtom)
 
   useEffect(() => {
     setIsLoading(true)
 
     async function getAllSecretsByProjectSlug() {
+      if (!selectedProject) {
+        toast.error('No project selected', {
+          description: (
+            <p className="text-xs text-red-300">
+              No project selected. Please select a project.
+            </p>
+          )
+        })
+        return
+      }
+
       const { success, error, data } =
         await ControllerInstance.getInstance().secretController.getAllSecretsOfProject(
-          { projectSlug: pathname.split('/')[2] },
+          { projectSlug: selectedProject.name },
           {}
         )
 
       if (success && data) {
-        setAllSecrets(data.items)
+        setSecrets(data.items)
       } else {
         // eslint-disable-next-line no-console -- we need to log the error
         console.error(error)
@@ -58,7 +56,7 @@ function SecretPage(): React.JSX.Element {
     getAllSecretsByProjectSlug()
 
     setIsLoading(false)
-  }, [pathname])
+  }, [selectedProject])
 
   if (isLoading) {
     return (
@@ -71,91 +69,40 @@ function SecretPage(): React.JSX.Element {
   }
 
   return (
-    <ScrollArea className=" mb-4 h-[50rem]">
-      <Accordion
-        className="flex h-[50rem] flex-col gap-4"
-        collapsible
-        type="single"
-      >
-        {allSecrets?.map(({ secret, values }) => {
-          return (
-            <AccordionItem
-              className="rounded-xl bg-white/5 px-5"
-              key={secret.id}
-              value={secret.id}
-            >
-              <AccordionTrigger
-                className="hover:no-underline"
-                rightChildren={
-                  <div className="text-xs text-white/50">
-                    {dayjs(secret.updatedAt).toNow(true)} ago by{' '}
-                    <span className="text-white">{secret.lastUpdatedById}</span>
-                  </div>
-                }
-              >
-                <div className="flex gap-x-5">
-                  <div className="flex items-center gap-x-4">
-                    {/* <SecretLogoSVG /> */}
-                    {secret.name}
-                  </div>
-                  {secret.note ? (
-                    <TooltipProvider>
-                      <Tooltip>
-                        <TooltipTrigger>
-                          <NoteIconSVG className="w-7" />
-                        </TooltipTrigger>
-                        <TooltipContent className="border-white/20 bg-white/10 text-white backdrop-blur-xl">
-                          <p>{secret.note}</p>
-                        </TooltipContent>
-                      </Tooltip>
-                    </TooltipProvider>
-                  ) : null}
-                </div>
-              </AccordionTrigger>
-              <AccordionContent>
-                <Table>
-                  <TableHeader>
-                    <TableRow>
-                      <TableHead>Environment</TableHead>
-                      <TableHead>Secret</TableHead>
-                    </TableRow>
-                  </TableHeader>
-                  <TableBody>
-                    {values.map((value) => {
-                      return (
-                        <TableRow key={value.environment.id}>
-                          <TableCell>{value.environment.slug}</TableCell>
-                          <TableCell className="max-w-40 overflow-auto">
-                            {value.value}
-                          </TableCell>
-                        </TableRow>
-                      )
-                    })}
-                  </TableBody>
-                </Table>
-              </AccordionContent>
-            </AccordionItem>
-          )
-        })}
-      </Accordion>
-    </ScrollArea>
-  )
-}
+    <div className={` flex h-full w-full justify-center `}>
+      {secrets.length === 0 ? (
+        <div className="flex h-[23.75rem] w-[30.25rem] flex-col items-center justify-center gap-y-8">
+          <SecretLogoSVG />
 
-function SecretLoader(): React.JSX.Element {
-  return (
-    <div className=" rounded-xl bg-white/5 p-4">
-      <div className="flex justify-between">
-        <div className="flex items-center gap-x-6">
-          <Skeleton className=" h-6 w-32 rounded" />
-          <Skeleton className=" size-6 rounded" />
+          <div className="flex h-[5rem] w-[30.25rem] flex-col items-center justify-center gap-4">
+            <p className="h-[2.5rem] w-[30.25rem] text-center text-[32px] font-[400]">
+              Declare your first secret
+            </p>
+            <p className="h-[1.5rem] w-[30.25rem] text-center text-[16px] font-[500]">
+              Declare and store a secret against different environments
+            </p>
+          </div>
+
+          <Button
+            className="h-[2.25rem] w-[8rem] rounded-md bg-white text-black hover:bg-gray-300"
+            onClick={() => setIsCreateSecretOpen(true)}
+          >
+            Create secret
+          </Button>
         </div>
-        <div className="flex items-center gap-x-3">
-          <Skeleton className=" h-6 w-24 rounded" />
-          <Skeleton className=" h-6 w-16 rounded" />
-          <Skeleton className=" ml-5 size-4 rounded" />
-        </div>
-      </div>
+      ) : (
+        <ScrollArea className=" mb-4 h-[50rem]">
+          <Accordion
+            className="flex h-[50rem] flex-col gap-4"
+            collapsible
+            type="single"
+          >
+            {secrets.map(({ secret, values }) => (
+              <SecretCard key={secret.id} {...{ secret, values }} />
+            ))}
+          </Accordion>
+        </ScrollArea>
+      )}
     </div>
   )
 }
